requirements.txt
asgiref==3.7.2
Django==4.2.3
email-validator==2.1.0.post1
flake8==7.0.0
flake8-bugbear==23.12.2
flake8-builtins==2.2.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-print==5.0.0
flake8-variables-names==0.0.6
isort==5.13.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
pycodestyle==2.11.1
pydantic==2.6.1
pyflakes==3.2.0
pytest_django==4.8.0
responses==0.24.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
